+++ custom-addons/cristal_agent/services/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from . import claude_client
 from . import prompt_builder
+from . import transcription
 from . import tool_registry
 from . import tools